+++ benchmark/benches/bench_world_new_entity_1_soa.jl
@@ -1,8 +1,8 @@
 
 function setup_world_new_entity_1_soa(n::Int)
     world = World(
-        Position => Storage{StructArray},
-        Velocity => Storage{StructArray};
+        Position => Storage(StructArray),
+        Velocity => Storage(StructArray);
         boxed=BOXED,
     )